requirements.txt
SQLAlchemy>=2.0
python-dotenv